convert to double

+++ src/activation.cu
@@ -1,15 +1,15 @@
 # include "activation.h"
 
 
-__device__ float activation_relu(float x) {
+__device__ double activation_relu(double x) {
     return x < 0 ? 0 : x;
 }
 
-float activation_relu_cpu(float x) {
+double activation_relu_cpu(double x) {
     return x < 0 ? 0 : x;
 }
 
-__global__ void forward_relu(float *input, float *output, const int size) {
+__global__ void forward_relu(double *input, double *output, const int size) {
     const int thread_pos = blockIdx.x * blockDim.x + threadIdx.x;
     const int total_threads = blockDim.x * gridDim.x;
     const int begin_idx = size * thread_pos / total_threads;
@@ -18,19 +18,19 @@ __global__ void forward_relu(float *input, float *output, const int size) {
         output[i] = activation_relu(input[i]);
 }
 
-float* cpu_relu(float *input, const int size, bool inplace) {
-    float *output;
+double* cpu_relu(double *input, const int size, bool inplace) {
+    double *output;
     if (inplace) output = input;
-    else output = (float*) malloc (sizeof(float) * size);
+    else output = (double*) malloc (sizeof(double) * size);
     for (int i = 0; i < size; ++ i)
         output[i] = activation_relu_cpu(input[i]);
     return output;
 }
 
-float* relu(dim3 grid, dim3 block, float *input, const int size, bool inplace) {
-    float *output;
+double* relu(dim3 grid, dim3 block, double *input, const int size, bool inplace) {
+    double *output;
     if (inplace) output = input;
-    else cudaMalloc((void **)&output, sizeof(float) * size);
+    else cudaMalloc((void **)&output, sizeof(double) * size);
     forward_relu <<<grid, block>>> (input, output, size);
     return output;
 }
