+++ src/opers.cu
@@ -1,7 +1,7 @@
 # include "opers.h"
 
 
-__global__ void forward_linear_transform(float *input, float *output, const int size, const float alpha, const float beta) {
+__global__ void forward_linear_transform(double *input, double *output, const int size, const double alpha, const double beta) {
     const int thread_pos = blockIdx.x * blockDim.x + threadIdx.x;
     const int total_threads = blockDim.x * gridDim.x;
     const int begin_idx = 1ll * size * thread_pos / total_threads;
@@ -10,28 +10,28 @@ __global__ void forward_linear_transform(float *input, float *output, const int 
         output[i] = input[i] * alpha + beta;
 }
 
-float* cpu_linear_transform(float *input, const int size, const float alpha, const float beta, bool inplace) {
-    float *output;
+double* cpu_linear_transform(double *input, const int size, const double alpha, const double beta, bool inplace) {
+    double *output;
     if (inplace) output = input;
-    else output = (float*) malloc (sizeof(float) * size);
+    else output = (double*) malloc (sizeof(double) * size);
     for (int i = 0; i < size; ++ i)
         output[i] = input[i] * alpha + beta;
     return output;
 }
 
-float* linear_transform(dim3 grid, dim3 block, float *input, const int size, const float alpha, const float beta, bool inplace) {
-    float *output;
+double* linear_transform(dim3 grid, dim3 block, double *input, const int size, const double alpha, const double beta, bool inplace) {
+    double *output;
     if (inplace) output = input;
-    else cudaMalloc((void **)&output, sizeof(float) * size);
+    else cudaMalloc((void **)&output, sizeof(double) * size);
     forward_linear_transform <<<grid, block>>> (input, output, size, alpha, beta);
     return output;
 }
 
-float* cpu_channel_concat(float *input[], const int num, const int batch_size, const int channel[], const int size_r, const int size_c) {
+double* cpu_channel_concat(double *input[], const int num, const int batch_size, const int channel[], const int size_r, const int size_c) {
     int total_channels = 0;
     for (int i = 0; i < num; ++ i) total_channels += channel[i];
-    float *output;
-    output = (float*) malloc (sizeof(float) * batch_size * total_channels * size_r * size_c);
+    double *output;
+    output = (double*) malloc (sizeof(double) * batch_size * total_channels * size_r * size_c);
     int cur_channels = 0;
     for (int i = 0; i < num; ++ i) {
         for (int b = 0; b < batch_size; ++ b)
@@ -44,7 +44,7 @@ float* cpu_channel_concat(float *input[], const int num, const int batch_size, c
     return output;
 }
 
-__global__ void forward_channel_concat_2(float *input1, float *input2, float *output, const int batch_size, const int channel1, const int channel2, const int size_r, const int size_c) {
+__global__ void forward_channel_concat_2(double *input1, double *input2, double *output, const int batch_size, const int channel1, const int channel2, const int size_r, const int size_c) {
     const int thread_pos = blockIdx.x * blockDim.x + threadIdx.x;
     const int batch_id = blockIdx.y;
     const int total_threads = blockDim.x * gridDim.x;
@@ -61,12 +61,12 @@ __global__ void forward_channel_concat_2(float *input1, float *input2, float *ou
         const int ch = (temp /= size_r) % total_channels;
         const int in_size = (ch < channel1 ? in_size_1 : in_size_2);
         const int cur_channels = (ch < channel1 ? 0 : channel1);
-        float *cur_input = (ch < channel1 ? input1 : input2);
+        double *cur_input = (ch < channel1 ? input1 : input2);
         output[batch_id * out_size + ch * size_r * size_c + r * size_c + c] = cur_input[batch_id * in_size + (ch - cur_channels) * size_r * size_c + r * size_c + c];
     }
 }
 
-__global__ void forward_channel_concat_3(float *input1, float *input2, float *input3, float *output, const int batch_size, const int channel1, const int channel2, const int channel3, const int size_r, const int size_c) {
+__global__ void forward_channel_concat_3(double *input1, double *input2, double *input3, double *output, const int batch_size, const int channel1, const int channel2, const int channel3, const int size_r, const int size_c) {
     const int thread_pos = blockIdx.x * blockDim.x + threadIdx.x;
     const int batch_id = blockIdx.y;
     const int total_threads = blockDim.x * gridDim.x;
@@ -84,12 +84,12 @@ __global__ void forward_channel_concat_3(float *input1, float *input2, float *in
         const int ch = (temp /= size_r) % total_channels;
         const int in_size = (ch < channel1 ? in_size_1 : ((ch < channel1 + channel2) ? in_size_2 : in_size_3));
         const int cur_channels = (ch < channel1 ? 0 : ((ch < channel1 + channel2) ? channel1 : (channel1 + channel2)));
-        float *cur_input = (ch < channel1 ? input1 : ((ch < channel1 + channel2) ? input2 : input3));
+        double *cur_input = (ch < channel1 ? input1 : ((ch < channel1 + channel2) ? input2 : input3));
         output[batch_id * out_size + ch * size_r * size_c + r * size_c + c] = cur_input[batch_id * in_size + (ch - cur_channels) * size_r * size_c + r * size_c + c];
     }
 }
 
-__global__ void forward_channel_concat_4(float *input1, float *input2, float *input3, float *input4, float *output, const int batch_size, const int channel1, const int channel2, const int channel3, const int channel4, const int size_r, const int size_c) {
+__global__ void forward_channel_concat_4(double *input1, double *input2, double *input3, double *input4, double *output, const int batch_size, const int channel1, const int channel2, const int channel3, const int channel4, const int size_r, const int size_c) {
     const int thread_pos = blockIdx.x * blockDim.x + threadIdx.x;
     const int batch_id = blockIdx.y;
     const int total_threads = blockDim.x * gridDim.x;
@@ -108,17 +108,17 @@ __global__ void forward_channel_concat_4(float *input1, float *input2, float *in
         const int ch = (temp /= size_r) % total_channels;
         const int in_size = (ch < channel1 ? in_size_1 : ((ch < channel1 + channel2) ? in_size_2 : ((ch < channel1 + channel2 + channel3) ? in_size_3 : in_size_4)));
         const int cur_channels = (ch < channel1 ? 0 : ((ch < channel1 + channel2) ? channel1 : ((ch < channel1 + channel2 + channel3) ? (channel1 + channel2) : (channel1 + channel2 + channel3))));
-        float *cur_input = (ch < channel1 ? input1 : ((ch < channel1 + channel2) ? input2 : ((ch < channel1 + channel2 + channel3) ? input3 : input4)));
+        double *cur_input = (ch < channel1 ? input1 : ((ch < channel1 + channel2) ? input2 : ((ch < channel1 + channel2 + channel3) ? input3 : input4)));
         output[batch_id * out_size + ch * size_r * size_c + r * size_c + c] = cur_input[batch_id * in_size + (ch - cur_channels) * size_r * size_c + r * size_c + c];
     }
 }
 
-float* channel_concat(dim3 grid, dim3 block, float *input[], const int num, const int batch_size, const int channel[], const int size_r, const int size_c) {
+double* channel_concat(dim3 grid, dim3 block, double *input[], const int num, const int batch_size, const int channel[], const int size_r, const int size_c) {
     assert(num >= 2 && num <= 4);
     int total_channels = 0;
     for (int i = 0; i < num; ++ i) total_channels += channel[i];
-    float *output;
-    cudaMalloc((void **)&output, sizeof(float) * batch_size * total_channels * size_r * size_c);
+    double *output;
+    cudaMalloc((void **)&output, sizeof(double) * batch_size * total_channels * size_r * size_c);
     if (num == 2) 
         forward_channel_concat_2 <<<grid, block>>> (input[0], input[1], output, batch_size, channel[0], channel[1], size_r, size_c);
     else if (num == 3)
