+++ src/pooling.h
@@ -5,11 +5,11 @@
 # include <stdio.h>
 # endif
 
-void maxpooling_cpu(float*, float*, int*, const int, const int, const int, const int, const int);
-void meanpooling_cpu(float*, float*, const int, const int, const int, const int, const int, const int);
+void maxpooling_cpu(double*, double*, int*, const int, const int, const int, const int, const int);
+void meanpooling_cpu(double*, double*, const int, const int, const int, const int, const int, const int);
 
-__global__ void maxpooling_forward(float*, float*, int*, const int, const int, const int);
-__global__ void meanpool_forward(float*, float*, const int, const int, const int, const int);
+__global__ void maxpooling_forward(double*, double*, int*, const int, const int, const int);
+__global__ void meanpool_forward(double*, double*, const int, const int, const int, const int);
 
 class MaxpoolingLayer{
     private:
@@ -17,8 +17,8 @@ class MaxpoolingLayer{
         int output_size;
     public:
         MaxpoolingLayer(int _channels, int _size, int _kernel_size, int _stride);
-        float* basic_forward(dim3 grid, dim3 block, float* input, const int batch_size);
-        float* cpu_forward(float *input, const int batch_size);
+        double* basic_forward(dim3 grid, dim3 block, double* input, const int batch_size);
+        double* cpu_forward(double *input, const int batch_size);
         ~MaxpoolingLayer();
 };
 
@@ -28,7 +28,7 @@ class MeanpoolingLayer{
         int output_size;
     public:
         MeanpoolingLayer(int _channels, int _size, int _kernel_size, int _stride, int _padding);
-        float* basic_forward(dim3 grid, dim3 block, float* input, const int batch_size);
-        float* cpu_forward(float *input, const int batch_size);
+        double* basic_forward(dim3 grid, dim3 block, double* input, const int batch_size);
+        double* cpu_forward(double *input, const int batch_size);
         ~MeanpoolingLayer();
 };
